Fix jbuilders not rendering
Okay so, it turns out you're actually not supposed to specify the json
option when using jbuilder, you just call render with a status and
that's it. Otherwise it just renders things as explicit json based on
whatever rails thinks it should be (probably just using `#to_json` on
whatever it's passed). Who knew!?