Add slugs to tasks for pretty URL's
This change adds an endpoint to the Tasks Controller that can find a
task by either its primary key or a slug, which is just a hyphenated
string based on its title. As an aside, this also cleans up an oversight
in the validations for title uniquness, and now both titles and slugs
are unique to tasks but scoped to their user, always, no conditionals.