Rename Task Completions to Task Assignments
It's just a way better name, not sure how I didn't think of it before.

Committing (heh) a grave sin by using `rename_table`. In any other
context this breaks things, but since I'm the only user of this app, I'm
going to forego the proper data migration practices in favor of just
getting this shipped.